requirements.txt
amqp==2.6.1
asgiref==3.4.1
attrs==21.2.0
autopep8==1.5.7
billiard==3.6.4.0
boto3==1.18.29
botocore==1.21.29
certifi==2021.5.30
charset-normalizer==2.0.7
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==5.5
cron-descriptor==1.4.0
dj-database-url==0.5.0
Django==3.2.6
django-celery-beat==2.1.0
django-cors-headers==3.11.0
django-crispy-forms==1.12.0
django-crontab==0.7.1
django-extensions==3.2.3
drf-spectacular==0.27.2
django-storages==1.11.1
django-taggit==1.5.1
django-timezone-field==4.2.3
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.3
influxdb-client==1.18.0
iniconfig==1.1.1
jmespath==0.10.0
kombu==4.6.11
pandas==2.0.3
packaging==20.9
paho-mqtt==1.5.1
pluggy==0.13.1
prompt-toolkit==3.0.43
psycopg2-binary==2.9.1
py==1.10.0
pycodestyle==2.7.0
pyparsing==2.4.7
pytest==6.2.4
pytest-cov==2.12.1
python-crontab==3.0.0
python-dateutil>=2.8.1,<3.0.0
python-dotenv==0.18.0
pytz==2021.1
redis==3.5.3
requests==2.26.0
Rx==3.2.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.26.5
vine==1.3.0
wcwidth==0.2.13
whitenoise==5.3.0
